requirements.txt
pixivpy3>=3.7
bilibili-api-python>=16
opennsfw2>=0.1
moviepy>=1.0.3
imageio[ffmpeg]>=2.0
Pillow>=10.0
pydantic-settings>=2.0
python-dotenv>=1.0
apscheduler>=3.10
loguru>=0.7
qrcode>=7.0
httpx>=0.25
numpy>=1.24
